requirements.txt
aniso8601==9.0.1
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.75.0
graphene==3.0
graphene-file-upload==1.3.0
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
h11==0.13.0
idna==3.3
promise==2.3
pyasn1==0.4.8
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.19.2
python-jose==3.3.0
rsa==4.8
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
starlette-graphene3==0.5.1
typing-extensions==4.1.1
uvicorn==0.17.6
